Moved Dev commands to a separate cog Updated some error handling messages
