+++ LastWarcraftLog.txt
@@ -1,1 +1,1 @@
-k6Wn714PAQwy8ZBq
+BxqHhJ8w1N6cX7M2
